Made the map javascript slightly more flexible and unobtrusive
Map was broken on /explore because the showAll var wasn't being set.

So I've refactored the code slightly so that now if the div containing
the map includes data-view="all" it'll show ALL plaques, otherwise
it'll search for plaques using the URL of the current page appended
with .json and the search params.